+++ Acuanet/Acuanet/Lectura.cs
@@ -9,13 +9,13 @@ namespace Acuanet
     //Clase para registra la lectura y el tiempo estima posicion y velocidad promedio
     class Lectura
     {
-        public double rssi;
+        public decimal rssi;
         public long tiempo;
         public int milis;
 
-        public double tms;
-        public double d_dist;
-        public double a_dist;
+        public decimal tms;
+        public decimal d_dist;
+        public decimal a_dist;
 
         public bool bdatoc=false;
     }
