+++ Acuanet/Acuanet/Resultado.cs
@@ -14,9 +14,9 @@ namespace Acuanet
         public int cantidad_aLec = 0;
 
         //ultima lectura
-        public double rssi_max;
-        public double tms_max;
-        public double d_min;
+        public decimal rssi_max;
+        public decimal tms_max;
+        public decimal d_min;
 
         //tiempo de cruce de la meta
         public decimal tc_meta;
